requirements.txt
gradio
mecab-python3
ipadic